+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.0-beta.10
 ### Added
 - Google address search functionality
+- Added google recaptcha v3
 
 ## 1.0.0-beta.8
 ### Changed